minskatt: initial commit
First commit of the project! :tada: :rocket:

We spent some time with @MalinMnnikk looking through the docs of
Skatteverket and trying to understand what is it that an SRU file really
consists of.

After that, we put together the initial grammar, and with this it was
pretty straightforward to hack together a LALRPOP parser in Rust that'll
chew a file into a typed representation of this info.

I took the liberty to implement a Display trait so printing stuff out is
easy, but there's still some work to do when it comes to:

1. validating the models
2. making a more robust parser

Oh yeah, and right now its a command line tool, which is probably less
than ideal.

But hey, at least it validates that this works! :D

More to come, stay tuned!